BVN64 fixed OSM render script
