+++ .github/workflows/outputs.yml
@@ -3,7 +3,7 @@ name: Generate and Persist Outputs
 on:
   push:
     branches:
-      - master
+      - '**'
 
 jobs:
   directory:
@@ -70,19 +70,12 @@ jobs:
       - name: Create Output Directory
         run: |
           mkdir -p ${{ steps.read_to_var.outputs.LOCAL_OUTPUT_DIR }}/main
-      - name: Update Satellite Imagery
-        run: |
-          cicd/imagery/scripts/update-network.sh
       - name: Export OSM XML
         run: |
           cicd/osm/scripts/export.sh local_output_dir=${{ steps.read_to_var.outputs.LOCAL_OUTPUT_DIR }}
       - name: Render OSM
         run: |
           cicd/osm/scripts/render.sh local_output_dir=${{ steps.read_to_var.outputs.LOCAL_OUTPUT_DIR }}
-      - name: Remove Unnecessary Files
-        run: |
-          sudo rm ${{ steps.read_to_var.outputs.LOCAL_OUTPUT_DIR }}/main/bvnordic.osm*.tif
-          sudo rm ${{ steps.read_to_var.outputs.LOCAL_OUTPUT_DIR }}/main/bvnordic.osm*.xml
       - name: Retain OSM Data
         uses: actions/upload-artifact@v2
         with:
