+++ cicd/imagery/scripts/update-all.sh
@@ -2,5 +2,4 @@
 
 pushd $(dirname $0)/../../..
 
-cicd/imagery/scripts/update-network-overview.sh
 cicd/imagery/scripts/update-stadium.sh
